+++ apps/client/src/environments/patch-notes.ts
@@ -1,9 +1,4 @@
 export const patchNotes = `### Bug Fixes
 
-* **db:** fixed coords for some nodes.
-* **db:** fixed db button not aligned properly for some nodes.
-
-
-### Features
-
-* **settings:** new import/export feature in general settings.`;
+* **core:** fixed major CPU usage issue caused by settings reloading system.
+* **updater:** fixed updating status not properly shown in topbar.`;